+++ history.md
@@ -9,3 +9,15 @@
 - January 4th: Started my map creation by cropping images for the grid
 - January 5th: Worked on my grid map level
 - January 6th: Finished my level 1 grid based map layout
+- January 8th: Added Ememy Class
+- January 10th: Worked on Enemy Class to follow player
+- January 13th: Found useful code online to finish Enemy Class to follow Cowboy (Credited inside sketch.js)
+- January 14th: Cleaned up code and removed things not needed
+- January 16th: Worked on starting a shot Class
+- January 17th: Worked on the shot Class
+- January 18th: Cleaned up and finished shot Class code
+- January 19th: Added collision to the Enemy and Cowboy
+	Added health and score system
+- January 20th: Added collision to the Shot from the Cowboy and Enemy
+	Revised health and score system
+- January 21st: Cleaned up and commented code